Move class to the shared lib
