+++ ide/che-core-ide-app/src/main/java/org/eclipse/che/ide/resources/impl/ProjectImpl.java
@@ -18,14 +18,17 @@ import com.google.common.base.MoreObjects;
 import com.google.common.base.Optional;
 import com.google.inject.Inject;
 import com.google.inject.assistedinject.Assisted;
+import java.util.Collections;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 import org.eclipse.che.api.core.model.project.ProjectConfig;
 import org.eclipse.che.api.core.model.project.ProjectProblem;
 import org.eclipse.che.api.core.model.project.SourceStorage;
 import org.eclipse.che.api.project.shared.dto.SourceEstimation;
 import org.eclipse.che.api.promises.client.Promise;
 import org.eclipse.che.api.promises.client.PromiseProvider;
+import org.eclipse.che.api.workspace.shared.ProjectProblemImpl;
 import org.eclipse.che.ide.api.resources.Project;
 import org.eclipse.che.ide.api.resources.marker.Marker;
 import org.eclipse.che.ide.resource.Path;
@@ -44,6 +47,7 @@ class ProjectImpl extends ContainerImpl implements Project {
   private static final int FOLDER_NOT_EXISTS_ON_FS = 10;
 
   private final ProjectConfig reference;
+  private final List<ProjectProblem> problems;
 
   @Inject
   protected ProjectImpl(
@@ -53,6 +57,16 @@ class ProjectImpl extends ContainerImpl implements Project {
     super(Path.valueOf(reference.getPath()), resourceManager, promiseProvider);
 
     this.reference = reference;
+    if (reference.getProblems() != null) {
+      problems =
+          reference
+              .getProblems()
+              .stream()
+              .map(problem -> new ProjectProblemImpl(problem.getCode(), problem.getMessage()))
+              .collect(Collectors.toList());
+    } else {
+      problems = Collections.emptyList();
+    }
   }
 
   /** {@inheritDoc} */
@@ -99,7 +113,7 @@ class ProjectImpl extends ContainerImpl implements Project {
 
   @Override
   public List<ProjectProblem> getProblems() {
-    return null;
+    return problems;
   }
 
   /** {@inheritDoc} */
